requirements/common.txt
azure-cognitiveservices-speech
aiohttp
sseclient-py
twitch-python
websockets
flask
pydouyu
pytchat
playsound
